+++ app/api/lessons/route.ts
@@ -1,13 +1,32 @@
 import { LessonService } from "@/service/lesson.service";
+import { getAuthenticatedUser } from "@/lib/auth";
+
 
 const lessonService = new LessonService();
-export async function GET(){
+export async function GET(request: Request){
+    const user = await getAuthenticatedUser(request);
+
+    if(user.role !== "ADMIN" && user.role !== "INSTRUCTOR") {
+        return Response.json(
+            { message: "Forbidden" },
+            { status: 403 }
+        )
+    }
     const lessons = await lessonService.getAllLessons();
 
     return Response.json(lessons);
 }
 
 export async function POST(request: Request) {
+    const user = await getAuthenticatedUser(request);
+
+    if(user.role !== "ADMIN" && user.role !== "INSTRUCTOR") {
+        return Response.json(
+            { message: "Forbidden" },
+            { status: 403 }
+        )
+    }
+
     const body = await request.json()
     const lesson = await lessonService.createLesson(body);
 
@@ -15,6 +34,15 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request){
+    const user = await getAuthenticatedUser(request);
+
+    if(user.role !== "ADMIN" && user.role !== "INSTRUCTOR") {
+        return Response.json(
+            { message: "Forbidden" },
+            { status: 403 }
+        )
+    }
+    
     const body = await request.json();
 
     const lesson = await lessonService.deleteLesson(body.id);
@@ -24,6 +52,15 @@ export async function DELETE(request: Request){
 
 
 export async function PATCH(request: Request){
+    const user = await getAuthenticatedUser(request);
+
+    if(user.role !== "ADMIN" && user.role !== "INSTRUCTOR") {
+        return Response.json(
+            { message: "Forbidden" },
+            { status: 403 }
+        )
+    }
+    
     const body = await request.json();
     const id = body.id;
     const lesson = await lessonService.updateLesson(id, body);
